server/requirements.txt
fastapi
uvicorn
openai
arxiv
pydantic
httpx

langchain
langchain-community
langchain-openai
python-dotenv